pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meetup-gmail-calendar-sync"
version = "0.1.0"
description = "Sync Meetup ICS invites from Gmail into Google Calendar"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Markster contributors"}
]
keywords = ["meetup", "gmail", "google-calendar", "sync", "ics"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Office/Business :: Scheduling",
]
dependencies = [
  "google-api-python-client>=2.170.0",
  "google-auth>=2.30.0",
  "google-auth-oauthlib>=1.2.0",
  "icalendar>=6.0.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.7.0",
]

[project.scripts]
meetup-gcal-sync = "meetup_gmail_calendar_sync.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
